Move rooms into Redux
