+++ src/features/board.ts
@@ -8,6 +8,30 @@ import { GridLoc } from '../components/board/grid';
 import { Direction } from '../components/room/Room';
 import { RootState } from '../rootReducer';
 
+export enum PlayerColor {
+  WHITE = 'white',
+  BLUE = 'blue',
+  GREEN = 'green',
+  YELLOW = 'yellow',
+  RED = 'red',
+  PURPLE = 'purple',
+}
+
+export interface Player {
+  color: PlayerColor;
+}
+
+export interface Room {
+  name: string;
+  loc: GridLoc;
+  doorDirections: Direction[];
+  players: Player[];
+}
+
+interface BoardState {
+  rooms: Room[];
+}
+
 const getMatchingDoor: (dir: Direction) => Direction = (dir) => {
   switch (dir) {
     case Direction.NORTH:
@@ -50,9 +74,45 @@ export const openSpotClicked: (
   dispatch(placeRoom({ loc }));
 };
 
-interface BoardState {}
-
-const initialState: BoardState = {};
+const initialState: BoardState = {
+  rooms: [
+    {
+      name: 'Bloody Room',
+      loc: { gridX: 2, gridY: 2 },
+      doorDirections: [
+        Direction.SOUTH,
+        Direction.EAST,
+        Direction.NORTH,
+        Direction.WEST,
+      ],
+      players: [],
+    },
+    {
+      name: 'Statuary Corridor',
+      loc: { gridX: 1, gridY: 2 },
+      doorDirections: [Direction.EAST, Direction.SOUTH],
+      players: [{ color: PlayerColor.BLUE }],
+    },
+    {
+      name: 'Master Bedroom',
+      loc: { gridX: 2, gridY: 3 },
+      doorDirections: [Direction.NORTH, Direction.SOUTH],
+      players: [],
+    },
+    {
+      name: 'Crypt',
+      loc: { gridX: 2, gridY: 1 },
+      doorDirections: [Direction.SOUTH],
+      players: [
+        { color: PlayerColor.YELLOW },
+        { color: PlayerColor.RED },
+        { color: PlayerColor.GREEN },
+        { color: PlayerColor.WHITE },
+        { color: PlayerColor.PURPLE },
+      ],
+    },
+  ],
+};
 
 const boardSlice = createSlice({ name: 'board', initialState, reducers: {} });
